+++ pybamm/models/base_model.py
@@ -115,11 +115,15 @@ class BaseModel(object):
         self._jacobian = None
         self._jacobian_algebraic = None
         self.external_variables = []
+        self._input_parameters = None
 
         # Default behaviour is to use the jacobian and simplify
         self.use_jacobian = True
         self.use_simplify = True
         self.convert_to_format = "casadi"
+
+        # Model is not initially discretised
+        self.is_discretised = False
 
         # Default timescale is 1 second
         self.timescale = pybamm.Scalar(1)
@@ -317,6 +321,25 @@ class BaseModel(object):
         "Set the timescale"
         self._timescale = value
 
+    @property
+    def input_parameters(self):
+        "Returns all the input parameters in the model"
+        if self._input_parameters is None:
+            self._input_parameters = self._find_input_parameters()
+        return self._input_parameters
+
+    def _find_input_parameters(self):
+        "Find all the input parameters in the model"
+        unpacker = pybamm.SymbolUnpacker(pybamm.InputParameter)
+        all_input_parameters = unpacker.unpack_list_of_symbols(
+            list(self.rhs.values())
+            + list(self.algebraic.values())
+            + list(self.initial_conditions.values())
+            + list(self.variables.values())
+            + [event.expression for event in self.events]
+        )
+        return list(all_input_parameters.values())
+
     def __getitem__(self, key):
         return self.rhs[key]
 
@@ -398,13 +421,13 @@ class BaseModel(object):
             for node in eq.pre_order():
                 if isinstance(node, pybamm.VariableDot):
                     raise pybamm.ModelError(
-                        "time derivative of variable"
-                        + " found ({}) in rhs equation {}".format(node, key)
+                        "time derivative of variable found "
+                        "({}) in rhs equation {}".format(node, key)
                     )
                 if isinstance(node, pybamm.StateVectorDot):
                     raise pybamm.ModelError(
-                        "time derivative of state vector"
-                        + " found ({}) in rhs equation {}".format(node, key)
+                        "time derivative of state vector found "
+                        "({}) in rhs equation {}".format(node, key)
                     )
 
         # Check that no variable time derivatives exist in the algebraic equations
@@ -433,46 +456,43 @@ class BaseModel(object):
         # For equations we look through the whole expression tree.
         # "Variables" can be Concatenations so we also have to look in the whole
         # expression tree
+        unpacker = pybamm.SymbolUnpacker((pybamm.Variable, pybamm.VariableDot))
+
         for var, eqn in self.rhs.items():
+            # Find all variables and variabledot objects
+            vars_in_rhs_keys_dict = unpacker.unpack_symbol(var)
+            vars_in_eqns_dict = unpacker.unpack_symbol(eqn)
+
+            # Store ids only
+            # Look only for Variable (not VariableDot) in rhs keys
             vars_in_rhs_keys.update(
-                [x.id for x in var.pre_order() if isinstance(x, pybamm.Variable)]
-            )
-            vars_in_eqns.update(
-                [x.id for x in eqn.pre_order() if isinstance(x, pybamm.Variable)]
-            )
-            vars_in_eqns.update(
                 [
-                    x.get_variable().id
-                    for x in eqn.pre_order()
-                    if isinstance(x, pybamm.VariableDot)
+                    var_id
+                    for var_id, var in vars_in_rhs_keys_dict.items()
+                    if isinstance(var, pybamm.Variable)
                 ]
             )
+            vars_in_eqns.update(vars_in_eqns_dict.keys())
         for var, eqn in self.algebraic.items():
+            # Find all variables and variabledot objects
+            vars_in_algebraic_keys_dict = unpacker.unpack_symbol(var)
+            vars_in_eqns_dict = unpacker.unpack_symbol(eqn)
+
+            # Store ids only
+            # Look only for Variable (not VariableDot) in algebraic keys
             vars_in_algebraic_keys.update(
-                [x.id for x in var.pre_order() if isinstance(x, pybamm.Variable)]
-            )
-            vars_in_eqns.update(
-                [x.id for x in eqn.pre_order() if isinstance(x, pybamm.Variable)]
-            )
-            vars_in_eqns.update(
                 [
-                    x.get_variable().id
-                    for x in eqn.pre_order()
-                    if isinstance(x, pybamm.VariableDot)
+                    var_id
+                    for var_id, var in vars_in_algebraic_keys_dict.items()
+                    if isinstance(var, pybamm.Variable)
                 ]
             )
+            vars_in_eqns.update(vars_in_eqns_dict.keys())
         for var, side_eqn in self.boundary_conditions.items():
             for side, (eqn, typ) in side_eqn.items():
-                vars_in_eqns.update(
-                    [x.id for x in eqn.pre_order() if isinstance(x, pybamm.Variable)]
-                )
-                vars_in_eqns.update(
-                    [
-                        x.get_variable().id
-                        for x in eqn.pre_order()
-                        if isinstance(x, pybamm.VariableDot)
-                    ]
-                )
+                vars_in_eqns_dict = unpacker.unpack_symbol(eqn)
+                vars_in_eqns.update(vars_in_eqns_dict.keys())
+
         # If any keys are repeated between rhs and algebraic then the model is
         # overdetermined
         if not set(vars_in_rhs_keys).isdisjoint(vars_in_algebraic_keys):
@@ -510,11 +530,12 @@ class BaseModel(object):
         equation
         """
         vars_in_bcs = set()
-        for var, side_eqn in self.boundary_conditions.items():
-            for eqn, _ in side_eqn.values():
-                vars_in_bcs.update(
-                    [x.id for x in eqn.pre_order() if isinstance(x, pybamm.Variable)]
-                )
+        unpacker = pybamm.SymbolUnpacker(pybamm.Variable)
+        for side_eqn in self.boundary_conditions.values():
+            all_vars = unpacker.unpack_list_of_symbols(
+                [eqn for eqn, _ in side_eqn.values()]
+            )
+            vars_in_bcs.update(all_vars.keys())
         if not post_discretisation:
             # After the model has been defined, each algebraic equation key should
             # appear in that algebraic equation, or in the boundary conditions
@@ -533,7 +554,7 @@ class BaseModel(object):
             # with the state vectors in the algebraic equations. Instead, we check
             # that each algebraic equation contains some StateVector
             for eqn in self.algebraic.values():
-                if not any(isinstance(x, pybamm.StateVector) for x in eqn.pre_order()):
+                if not eqn.has_symbol_of_classes(pybamm.StateVector):
                     raise pybamm.ModelError(
                         "each algebraic equation must contain at least one StateVector"
                     )
@@ -562,12 +583,8 @@ class BaseModel(object):
                     for x in symbol.pre_order()
                 ):
                     raise pybamm.ModelError(
-                        """
-                        no boundary condition given for
-                        variable '{}' with equation '{}'.
-                        """.format(
-                            var, eqn
-                        )
+                        "no boundary condition given for "
+                        "variable '{}' with equation '{}'.".format(var, eqn)
                     )
 
     def check_default_variables_dictionaries(self):
@@ -590,12 +607,9 @@ class BaseModel(object):
 
     def check_variables(self):
         # Create list of all Variable nodes that appear in the model's list of variables
-        all_vars = {}
-        for eqn in self.variables.values():
-            # Add all variables in the equation to the list of variables
-            all_vars.update(
-                {x.id: x for x in eqn.pre_order() if isinstance(x, pybamm.Variable)}
-            )
+        unpacker = pybamm.SymbolUnpacker(pybamm.Variable)
+        all_vars = unpacker.unpack_list_of_symbols(self.variables.values())
+
         var_ids_in_keys = set()
 
         model_and_external_variables = (
@@ -648,6 +662,26 @@ class BaseModel(object):
             symbol.print_input_names()
 
         print(div)
+
+    @property
+    def default_parameter_values(self):
+        return pybamm.ParameterValues({})
+
+    @property
+    def default_var_pts(self):
+        return {}
+
+    @property
+    def default_geometry(self):
+        return {}
+
+    @property
+    def default_submesh_types(self):
+        return {}
+
+    @property
+    def default_spatial_methods(self):
+        return {}
 
     @property
     def default_solver(self):