+++ pybamm/models/full_battery_models/lead_acid/__init__.py
@@ -3,5 +3,12 @@
 #
 from .base_lead_acid_model import BaseModel
 from .loqs import LOQS
-from .higher_order import BaseHigherOrderModel, FOQS, Composite, CompositeExtended
+from .higher_order import (
+    BaseHigherOrderModel,
+    FOQS,
+    Composite,
+    CompositeAverageCorrection,
+    CompositeExtended,
+)
 from .full import Full
+from .basic_full import BasicFull