+++ pybamm/models/full_battery_models/lithium_ion/spm.py
@@ -58,7 +58,12 @@ class SPM(BaseModel):
 
     def set_convection_submodel(self):
 
-        self.submodels["convection"] = pybamm.convection.NoConvection(self.param)
+        self.submodels[
+            "through-cell convection"
+        ] = pybamm.convection.through_cell.NoConvection(self.param)
+        self.submodels[
+            "transverse convection"
+        ] = pybamm.convection.transverse.NoConvection(self.param)
 
     def set_interfacial_submodel(self):
 
@@ -109,13 +114,12 @@ class SPM(BaseModel):
 
     def set_electrolyte_submodel(self):
 
-        electrolyte = pybamm.electrolyte.stefan_maxwell
-        surf_form = electrolyte.conductivity.surface_potential_form
+        surf_form = pybamm.electrolyte_conductivity.surface_potential_form
 
         if self.options["surface form"] is False:
             self.submodels[
                 "leading-order electrolyte conductivity"
-            ] = electrolyte.conductivity.LeadingOrder(self.param)
+            ] = pybamm.electrolyte_conductivity.LeadingOrder(self.param)
 
         elif self.options["surface form"] == "differential":
             for domain in ["Negative", "Separator", "Positive"]:
@@ -132,7 +136,7 @@ class SPM(BaseModel):
                 ] = surf_form.LeadingOrderAlgebraic(self.param, domain, self.reactions)
         self.submodels[
             "electrolyte diffusion"
-        ] = electrolyte.diffusion.ConstantConcentration(self.param)
+        ] = pybamm.electrolyte_diffusion.ConstantConcentration(self.param)
 
     def set_anode_decomposition_submodel(self):