+++ pybamm/models/submodels/electrolyte/__init__.py
@@ -1,4 +1,0 @@
-from .base_electrolyte_conductivity import BaseElectrolyteConductivity
-from .base_electrolyte_diffusion import BaseElectrolyteDiffusion
-
-from . import stefan_maxwell