+++ pybamm/models/submodels/electrolyte/stefan_maxwell/diffusion/__init__.py
@@ -1,6 +1,0 @@
-from .base_stefan_maxwell_diffusion import BaseModel
-from .leading_stefan_maxwell_diffusion import LeadingOrder
-from .first_order_stefan_maxwell_diffusion import FirstOrder
-from .composite_stefan_maxwell_diffusion import Composite
-from .full_stefan_maxwell_diffusion import Full
-from .constant_stefan_maxwell_diffusion import ConstantConcentration