+++ pybamm/parameters/parameter_sets.py
@@ -54,17 +54,16 @@ Chen2020 = {
     "citation": "Chen2020",
 }
 
-Cai2019 = { #same as Chen2020 but includes additional Cai2019 thermal runaway parameters
+Mohtat2020 = {
     "chemistry": "lithium-ion",
-    "cell": "LGM50_Cai2019",
-    "anode": "graphite_Cai2019",
-    "separator": "separator_Chen2020",
-    "cathode": "nmc_Cai2019",
-    "electrolyte": "lipf6_Nyman2008",
-    "experiment": "1C_discharge_from_full_Chen2020",
-    "citation": ["Chen2020", "cai2019modeling"], 
+    "cell": "UMBL_Mohtat2020",
+    "anode": "graphite_UMBL_Mohtat2020",
+    "separator": "separator_Mohtat2020",
+    "cathode": "NMC_UMBL_Mohtat2020",
+    "electrolyte": "LiPF6_Mohtat2020",
+    "experiment": "1C_charge_from_empty_Mohtat2020",
+    "citation": "Mohtat2020",
 }
-
 #
 # Lead-acid
 #