+++ tests/unit/test_discretisations/test_discretisation.py
@@ -705,8 +705,11 @@ class TestDiscretise(unittest.TestCase):
         # test that any time derivatives of variables in rhs raises an
         # error
         model = pybamm.BaseModel()
-        model.rhs = {c: pybamm.div(N) + c.diff(pybamm.t),
-                     T: pybamm.div(q), S: pybamm.div(p)}
+        model.rhs = {
+            c: pybamm.div(N) + c.diff(pybamm.t),
+            T: pybamm.div(q),
+            S: pybamm.div(p),
+        }
         model.initial_conditions = {
             c: pybamm.Scalar(2),
             T: pybamm.Scalar(5),
@@ -1013,6 +1016,7 @@ class TestDiscretise(unittest.TestCase):
             broad_disc.shape,
             (mesh["negative particle"][0].npts * mesh["negative electrode"][0].npts, 1),
         )
+        broad = pybamm.SecondaryBroadcast(var, "negative electrode")
 
         # test broadcast to edges
         broad_to_edges = pybamm.SecondaryBroadcastToEdges(var, "negative electrode")
@@ -1184,6 +1188,18 @@ class TestDiscretise(unittest.TestCase):
         np.testing.assert_equal(
             model.mass_matrix_inv.entries.toarray(), mass_inv.toarray()
         )
+
+    def test_process_input_variable(self):
+        disc = get_discretisation_for_testing()
+
+        a = pybamm.InputParameter("a")
+        a_disc = disc.process_symbol(a)
+        self.assertEqual(a_disc._expected_size, 1)
+
+        a = pybamm.InputParameter("a", ["negative electrode", "separator"])
+        a_disc = disc.process_symbol(a)
+        n = disc.mesh.combine_submeshes(*a.domain)[0].npts
+        self.assertEqual(a_disc._expected_size, n)
 
 
 if __name__ == "__main__":