+++ tests/unit/test_models/test_base_model.py
@@ -107,6 +107,32 @@ class TestBaseModel(unittest.TestCase):
         self.assertEqual(model[key], rhs[key])
         self.assertEqual(model[key], model.rhs[key])
 
+    def test_read_input_parameters(self):
+        # Read input parameters from different parts of the model
+        model = pybamm.BaseModel()
+        a = pybamm.InputParameter("a")
+        b = pybamm.InputParameter("b")
+        c = pybamm.InputParameter("c")
+        d = pybamm.InputParameter("d")
+        e = pybamm.InputParameter("e")
+        f = pybamm.InputParameter("f")
+
+        u = pybamm.Variable("u")
+        v = pybamm.Variable("v")
+        model.rhs = {u: -u * a}
+        model.algebraic = {v: v - b}
+        model.initial_conditions = {u: c, v: d}
+        model.events = [pybamm.Event("u=e", u - e)]
+        model.variables = {"v+f": v + f}
+
+        self.assertEqual(
+            set([x.name for x in model.input_parameters]),
+            set([x.name for x in [a, b, c, d, e, f]]),
+        )
+        self.assertTrue(
+            all(isinstance(x, pybamm.InputParameter) for x in model.input_parameters),
+        )
+
     def test_update(self):
         # model
         whole_cell = ["negative electrode", "separator", "positive electrode"]
@@ -260,10 +286,7 @@ class TestBaseModel(unittest.TestCase):
 
         # model must be in semi-explicit form
         model = pybamm.BaseModel()
-        model.algebraic = {
-            c: 2 * d - c,
-            d: c * d.diff(pybamm.t) - d,
-        }
+        model.algebraic = {c: 2 * d - c, d: c * d.diff(pybamm.t) - d}
         model.initial_conditions = {c: 1, d: 1}
         with self.assertRaisesRegex(
             pybamm.ModelError, "time derivative of variable found",