+++ tests/unit/test_simulation.py
@@ -86,6 +86,23 @@ class TestSimulation(unittest.TestCase):
             if val.size > 1:
                 self.assertTrue(val.has_symbol_of_classes(pybamm.Matrix))
 
+    def test_solve_non_battery_model(self):
+
+        model = pybamm.BaseModel()
+        v = pybamm.Variable("v")
+        model.rhs = {v: -v}
+        model.initial_conditions = {v: 1}
+        model.variables = {"v": v}
+        sim = pybamm.Simulation(
+            model, solver=pybamm.ScipySolver(rtol=1e-10, atol=1e-10)
+        )
+
+        sim.solve()
+        np.testing.assert_array_equal(sim.solution.t, np.linspace(0, 1, 100))
+        np.testing.assert_array_almost_equal(
+            sim.solution["v"].entries, np.exp(-np.linspace(0, 1, 100))
+        )
+
     def test_reuse_commands(self):
 
         sim = pybamm.Simulation(pybamm.lithium_ion.SPM())
@@ -146,10 +163,14 @@ class TestSimulation(unittest.TestCase):
         sim.build()
 
     def test_set_crate(self):
-        sim = pybamm.Simulation(pybamm.lithium_ion.SPM(), C_rate=2)
-        self.assertEqual(sim.parameter_values["C-rate"], 2)
+        model = pybamm.lithium_ion.SPM()
+        current_1C = model.default_parameter_values["Current function [A]"]
+        sim = pybamm.Simulation(model, C_rate=2)
+        self.assertEqual(sim.parameter_values["Current function [A]"], 2 * current_1C)
+        self.assertEqual(sim.C_rate, 2)
         sim.specs(C_rate=3)
-        self.assertEqual(sim.parameter_values["C-rate"], 3)
+        self.assertEqual(sim.parameter_values["Current function [A]"], 3 * current_1C)
+        self.assertEqual(sim.C_rate, 3)
 
     def test_set_defaults(self):
         sim = pybamm.Simulation(pybamm.lithium_ion.SPM())
@@ -193,7 +214,7 @@ class TestSimulation(unittest.TestCase):
         self.assertIsInstance(c_e, np.ndarray)
 
     def test_set_external_variable(self):
-        model_options = {"thermal": "x-lumped", "external submodels": ["thermal"]}
+        model_options = {"thermal": "lumped", "external submodels": ["thermal"]}
         model = pybamm.lithium_ion.SPMe(model_options)
         sim = pybamm.Simulation(model)
 
@@ -201,7 +222,7 @@ class TestSimulation(unittest.TestCase):
 
         dt = 0.001
 
-        external_variables = {"X-averaged cell temperature": T_av}
+        external_variables = {"Volume-averaged cell temperature": T_av}
         sim.step(dt, external_variables=external_variables)
 
     def test_step(self):
@@ -210,7 +231,7 @@ class TestSimulation(unittest.TestCase):
         model = pybamm.lithium_ion.SPM()
         sim = pybamm.Simulation(model)
         sim.step(dt)  # 1 step stores first two points
-        tau = model.timescale.evaluate()
+        tau = sim.model.timescale.evaluate()
         self.assertEqual(sim.solution.t.size, 2)
         self.assertEqual(sim.solution.y[0, :].size, 2)
         self.assertEqual(sim.solution.t[0], 0)
@@ -227,6 +248,14 @@ class TestSimulation(unittest.TestCase):
         self.assertEqual(sim.solution.t[0], 2 * dt / tau)
         self.assertEqual(sim.solution.t[1], 3 * dt / tau)
 
+    def test_solve_with_inputs(self):
+        model = pybamm.lithium_ion.SPM()
+        param = model.default_parameter_values
+        param.update({"Current function [A]": "[input]"})
+        sim = pybamm.Simulation(model, parameter_values=param)
+        sim.solve(inputs={"Current function [A]": 1})
+        np.testing.assert_array_equal(sim.solution.inputs["Current function [A]"], 1)
+
     def test_step_with_inputs(self):
         dt = 0.001
         model = pybamm.lithium_ion.SPM()
@@ -236,7 +265,7 @@ class TestSimulation(unittest.TestCase):
         sim.step(
             dt, inputs={"Current function [A]": 1}
         )  # 1 step stores first two points
-        tau = model.timescale.evaluate()
+        tau = sim.model.timescale.evaluate()
         self.assertEqual(sim.solution.t.size, 2)
         self.assertEqual(sim.solution.y[0, :].size, 2)
         self.assertEqual(sim.solution.t[0], 0)
@@ -313,13 +342,13 @@ class TestSimulation(unittest.TestCase):
         # make simulation with silly options (should this be allowed?)
         sim = pybamm.Simulation(
             model,
-            geometry=1,
-            parameter_values=1,
-            submesh_types=1,
-            var_pts=1,
-            spatial_methods=1,
-            solver=1,
-            quick_plot_vars=1,
+            geometry={},
+            parameter_values={},
+            submesh_types={},
+            var_pts={},
+            spatial_methods={},
+            solver={},
+            quick_plot_vars=[],
         )
 
         # reset and check
@@ -371,7 +400,7 @@ class TestSimulation(unittest.TestCase):
 
         # check solution is returned at the times in the data
         sim.solve()
-        tau = model.timescale.evaluate()
+        tau = sim.model.timescale.evaluate()
         np.testing.assert_array_almost_equal(sim.solution.t, time_data / tau)
 
         # check warning raised if the largest gap in t_eval is bigger than the
@@ -405,7 +434,7 @@ class TestSimulation(unittest.TestCase):
         )
         sim.solve()
         current = sim.solution["Current [A]"]
-        tau = model.timescale.evaluate()
+        tau = sim.model.timescale.evaluate()
         self.assertEqual(current(0), 1)
         self.assertEqual(current(1500 / tau), -0.5)
         self.assertEqual(current(3000 / tau), 0.5)